% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLS_Score.R
\name{CalTLSfea}
\alias{CalTLSfea}
\title{Calculation of TLS scores. TLS features stored in}
\usage{
CalTLSfea(
  data,
  st_pos,
  cluster,
  cutoff = 0.2,
  filt.dist = 4,
  filt.spots = 2,
  ...
)
}
\arguments{
\item{data}{A data.frame contains two TLS features, including
co-local scores of "Plasma/B.cells_T.cells" and enrichment scores of TLS
signature "LC.50sig".}

\item{st_pos}{A data.frame with coordinate information in two columns.}

\item{cluster}{Cluster labels of spots.}

\item{...}{Arguments passed to other methods.}
}
\value{
Lists containing distinct forms of TLS features and the final TLS scores.
}
\description{
Calculation of TLS scores. TLS features stored in
}
\seealso{
[RegFeaEnrich]
}
