% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit.R
\name{FindNeispots}
\alias{FindNeispots}
\title{Internal function to find neighbors for each spot within given clusters}
\usage{
FindNeispots(
  st_pos,
  cluster = NULL,
  method = "dist",
  r.dist = 2,
  p = NULL,
  return.list = T
)
}
\arguments{
\item{st_pos}{A data.frame with coordinate information in two columns.}

\item{cluster}{Cluster labels of spots.}

\item{r.dist}{A numeric value. Distance to define the neighboring spots. Default
is 2.}

\item{return.list}{Logical value. If TURE, return the neighbors of each spot
using list.}
}
\value{
List or matrix.
}
\description{
Internal function to find neighbors for each spot within given clusters
}
