% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit.R
\name{KernelWeight}
\alias{KernelWeight}
\title{Internal function to get weights based on distance.}
\usage{
KernelWeight(d, adjust = 1)
}
\arguments{
\item{d}{A numeric vector of distance.}

\item{adjust}{Adjustment of weights.}
}
\value{
A vector of weights named with distance (d).
}
\description{
Internal function to get weights based on distance.
}
