% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TN_interface.R
\name{TNIClass}
\alias{TNIClass}
\title{Categories of TNI regions}
\usage{
TNIClass(type, cluster, ES)
}
\arguments{
\item{type}{A named vector of type labels.}

\item{cluster}{A named vector of cluster.}

\item{ES}{Tumor cell abundances or other features representing the enrichment
of tumor cells.}
}
\value{
A vector of TNI categories named with spots.
}
\description{
Categories of TNI regions
}
